[XHAB-32] Renaming classes.
