+++ lib/src/presentation/scenes/auth/auth_state.dart
